feat: parts: filler: Add input to paste PEM certificates
The paste event on the 'cert' component wouldn't fire
all the time. Haven't been able to find reason for the
same.

As workaround, provide a readonly text area that allows
paste event. Allow paste event to propagate for parsing
the pasted contents.

Also add input to clear the text area. It shall be used
when pasted content is not a valid certificate.

Signed-off-by: Sanjeev Premi <spremi@ymail.com>